Revert linked lists merged into develop_p5. Will work on debugging before merging with rc_p5

+++ src-gui/p4/P4SepDlg.cpp
@@ -271,7 +271,7 @@ void P4SepDlg::onepsilon_enter()
     }
 
     gCurrentSeparatriceEpsilon = eps;
-    s.sprintf("%g", (float)eps);
+    s.sprintf("%g", static_cast<double>(eps));
     edt_epsilon_->setText(s);
 
     // pass on to math routines
